requirements.txt
streamlit==1.22.0
python-dotenv==1.0.0
unstructured==0.9.2
faiss-cpu==1.7.4
libmagic==1.0
python-magic==0.4.27
python-magic-bin==0.4.14

# Replace old LangChain
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.21
langchain-google-genai==0.0.9